Issue #566, #567 - automatic update webmail config when access to 'general settings', 'remote mail', 'webmail application' and etc

+++ kloxo/httpdocs/lib/serverweb__apachelib.php
@@ -16,11 +16,14 @@ function dbactionUpdate($subaction)
 
 	//-- old structure
 	passthru("rm -rf /etc/httpd/conf/kloxo");
-
+	passthru("rm -rf /home/httpd/conf");
+	
 	//-- new structure	
-	lxfile_mkdir("/home/httpd/conf");
-	lxfile_mkdir("/home/httpd/conf/defaults");
-	lxfile_mkdir("/home/httpd/conf/domains");
+	lxfile_mkdir("/home/apache/conf");
+	lxfile_mkdir("/home/apache/conf/defaults");
+	lxfile_mkdir("/home/apache/conf/domains");
+	lxfile_mkdir("/home/apache/conf/redirects");
+	lxfile_mkdir("/home/apache/conf/webmails");
 
 	//--- some vps include /etc/httpd/conf.d/swtune.conf
 	passthru("rm -f /etc/httpd/conf.d/swtune.conf");
@@ -28,7 +31,6 @@ function dbactionUpdate($subaction)
 	if (!lfile_exists("/etc/httpd/conf.d/~lxcenter.conf")) {
 		copy("/usr/local/lxlabs/kloxo/file/apache/~lxcenter.conf", "/etc/httpd/conf.d/~lxcenter.conf");
 		copy("/usr/local/lxlabs/kloxo/file/centos-5/httpd.conf", "/etc/httpd/conf/httpd.conf");
-
 	}
 
 	lxfile_rm("/etc/sysconfig/httpd");
@@ -38,6 +40,28 @@ function dbactionUpdate($subaction)
 	$a = $this->main->apache_optimize;
 	$m = $this->main->mysql_convert;
 	$f = $this->main->fix_chownchmod;
+
+	if ($f === 'fix-ownership') {
+		passthru("lphp.exe /usr/local/lxlabs/kloxo/bin/fix/fix-chownchmod.php --select=chown");
+	//	setFixChownChmod('chown');
+	}
+	else if ($f === 'fix-permissions') {
+		passthru("lphp.exe /usr/local/lxlabs/kloxo/bin/fix/fix-chownchmod.php --select=chmod");
+	//	setFixChownChmod('chmod');
+	}
+	else if ($f === 'fix-ALL') {
+		passthru("lphp.exe /usr/local/lxlabs/kloxo/bin/fix/fix-chownchmod.php --select=all");
+	//	setFixChownChmod('all');
+	}
+	
+	if ($m === 'to-myisam') {
+		passthru("lphp.exe /usr/local/lxlabs/kloxo/bin/fix/mysql-convert.php --engine=myisam");
+	//	setMysqlConvert('myisam');
+	}
+	else if ($m === 'to-innodb') {
+		passthru("lphp.exe /usr/local/lxlabs/kloxo/bin/fix/mysql-convert.php --engine=innodb");
+	//	setMysqlConvert('innodb');
+	}
 
 	//--- don't use '=== true' but '!== false'
 	if (strpos($t, 'mod_php') !== false) {
@@ -57,16 +81,21 @@ function dbactionUpdate($subaction)
 		}
 		else if ($t === 'mod_php_ruid2') {
 			passthru("yum -y install mod_ruid2");
+			passthru("yum -y update mod_ruid2");
 			lxfile_mv("/etc/httpd/conf.d/ruid2.nonconf", "/etc/httpd/conf.d/ruid2.conf");
 		}
 		else if ($t === 'mod_php_itk') {
-			passthru("yum -y install httpd-itk");
-			lxfile_rm("/etc/httpd/conf.d/itk.conf");
+			if (!lxfile_exists("/usr/sbin/httpd.itk")) {
+				passthru("yum -y install httpd-itk");
+				passthru("yum -y update httpd-itk");
+				lxfile_rm("/etc/httpd/conf.d/itk.conf");
+			}
 			passthru("echo 'HTTPD=/usr/sbin/httpd.itk' >/etc/sysconfig/httpd");
 		}
 	}
 	else if (strpos($t, 'suphp') !== false) {
 		passthru("yum -y install mod_suphp");
+		passthru("yum -y update mod_suphp");
 
 		lxfile_mv("/etc/httpd/conf.d/php.conf", "/etc/httpd/conf.d/php.nonconf");
 		lxfile_mv("/etc/httpd/conf.d/fastcgi.conf", "/etc/httpd/conf.d/fastgi.nonconf");
@@ -103,27 +132,10 @@ function dbactionUpdate($subaction)
 //	createRestartFile("httpd");
 
 //	lxshell_return("service", "httpd", "start");
-// 	passthru("/etc/init.d/httpd start");
+//	passthru("/etc/init.d/httpd start");
 
 	$ret = lxshell_return("service", "httpd", "start");
 	if ($ret) { throw new lxexception('httpd_start_failed', 'parent'); }
-
-	if ($f === 'fix-ownership') {
-		passthru("lphp.exe /usr/local/lxlabs/kloxo/bin/fix/fix-chownchmod.php --select=chown");
-	}
-	else if ($f === 'fix-permissions') {
-		passthru("lphp.exe /usr/local/lxlabs/kloxo/bin/fix/fix-chownchmod.php --select=chmod");
-	}
-	else if ($f === 'fix-ALL') {
-		passthru("lphp.exe /usr/local/lxlabs/kloxo/bin/fix/fix-chownchmod.php --select=all");
-	}
-	
-	if ($m === 'to-myisam') {
-		passthru("lphp.exe /usr/local/lxlabs/kloxo/bin/fix/mysql-convert.php --engine=myisam");
-	}
-	else if ($m === 'to-innodb') {
-		passthru("lphp.exe /usr/local/lxlabs/kloxo/bin/fix/mysql-convert.php --engine=innodb");
-	}
 
 	if ($a === 'optimize') {
 		passthru("lphp.exe /usr/local/lxlabs/kloxo/bin/fix/apache-optimize.php --select=optimize");
